Issue #62 Make OffHeapResource a CommonComponent
When performing a org.terracotta.entity.ServiceRegistry.getService call
against OffHeapResource, the configured resources can not be found
due to a difference in the ClassLoaders being used.  Adding the
@CommonComponent annotation to OffHeapResource causes the common
ClassLoader to be used for OffHeapResource so the service-created
instances can be found.  Since OffHeapResourceIdentifier class
references are also used during OffHeapResource lookup (instanceof),
this class is also marked with @CommonComponent.

This commit separates the OffHeapResource interface from its
implementation to avoid an issue during instantiation after application
of the @CommonComponent annotation.
